Add Babel ES6 transpiler to webpack
This configures the Babel ES6 to ES5 transpiler for JS files.

Turns out the JSX transformer already supports certain ES6 features using
its --harmony flag. I'm not sure what it supports exactly (seems to be
poorly documented), but at least it knows arrows. And we already had it
enabled.

For regular JS files we could of course also apply the JSX transformer,
but I don't think that's a good idea. Instead we use Babel for this.