Use ESLint instead of JSHint and JSXHint
Included .eslintrc contains a somewhat arbitrary set of rules and
especially their severity is just picked like that. It is based
on [1], extended according to [2] and [3], and then hand-modified
using the ESLint rules docs and current state of our code.

Lot's of violations currently, will be fixed in a future commit.

[1] https://raw.githubusercontent.com/clark800/javascript-style-guide/master/eslintrc
[2] https://github.com/eslint/eslint/issues/1669#issuecomment-70459612
[3] https://github.com/purposeindustries/styleguide/blob/master/.eslintrc
[4] http://eslint.org/docs/rules/

Fixes pedigree/pedigree#23