Simple reading of Excel files
We read Excel files following a predefined structure (including column
headers) as shown in the examples/example.xlsx file.

The xlsx module we use supports .xlsx files (not .xls) and .ods (not
implemented here).

Note that we skip packing the bundled jszip.js as it is not a source
module to get rid of the following Webpack warning:

> This seem to be a pre-built javascript file. Even while this is
> possible, it's not recommended. Try to require to orginal source to
> get better results.

Also note that by including xlsx we doubled our packed distribution
file (700K to 1.4M).

[1] https://github.com/SheetJS/js-xlsx