Rename zygotes to children and replace monozygotic by zygotes
Following the discussion in #34, zygosity within a pregnancy is
implemented using an explicit child-zygosity mapping as follows.

A pregnancy has a list of children (member keys) and a list of
corresponding zygosity numbers (integers). The zygosity numbers
could be of any type that is comparable, but we require them to be
integers for simplicity.

The special *unknown zygosity* state is represented by setting the
list of zygosity numbers to `undefined` (in our application model)
or by omitting it (in JSON).

Drawing is for now limited to complete homozygous pregnancies,
unknown zygosity, and all separate zygotes. All other zygosity
states are treated as unknown zygosity (drawn with question mark).

We do not yet support changing zygosity or setting the zygosity
for new pregnancies or twins.

In principle, this representation also give the possibility of
having an *unknown zygosity* state on the child level instead of
on the pregnancy level, by using `undefined` as a value in the
list of zygosity numbers. We do not allow this for now, since the
pedigree nomenclature does not show us how to draw this.

Fixes #34
Fixes #30