Add and remove custom fields
The form (for adding a custom field) is a bit rough, but I suspect
we will use plexus-forms for that later.

To ease generalization of functionality over our different object
types (pedigree, nest, member), the `FocusLevel` constant is renamed
to `ObjectType` and used to parameterize by object type. Similarly,
the `Focus` record is generalized to `Ref` as it is useful for more
than storing the focussed object.