Example website without IFrame
Considering some of the hard-to-debug issues we experience related to
the IFrame, we switch to a non-iframe development environment.

For now, the non-iframe environment is our primary environment, which
should support most of our medium-term goals. The iframe environment
is still available, but secondary. We'll revisit its bugs when we
start embedding the editor in 3rd party websites.