Flatten pedigree data model to tabular structure
This is a huge change touching nearly all of the application. The
internal data model used for pedigree structures is now similar to what
is used by Madeline.

The original data model used separate 'nest' entities in addition to
'member' entities. The new data model is more similar to a simple table
such as used by PED, with only 'member' entities linked by 'father' and
'mother' columns. We loose some more complex functionality with this,
but the integration with Madeline is much easier.

For now, the FAM parser is not implemented.